generate: Split out generate package for import
