+++ generate/generate_test.go
@@ -1,4 +1,4 @@
-package main
+package generate
 
 import (
 	"testing"
